Update navbar color scheme in CSS variables
Added a new color variable '--color-evergreen-dark' with a darker shade of green. Updated the navbar's background, font color, and hover background color to improve contrast and accessibility. Now, the navbar background is '--color-evergreen', the font color is '--color-white', and the hover background color is '--color-evergreen-dark'.

Release-Note: Update navbar color scheme for improved accessibility

Signed-off-by: IanSkelskey <ianskelskey@gmail.com>